Fix unsafe narrowing instances in ui/base.
This requires touching base/ and net/ also.

Bug: 1292951
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3827415
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Ryan Hamilton <rch@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1036681}

+++ net/filter/gzip_header.h
@@ -16,6 +16,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #ifndef NET_FILTER_GZIP_HEADER_H_
 #define NET_FILTER_GZIP_HEADER_H_
 
+#include <stddef.h>
 #include <stdint.h>
 
 #include "net/base/net_export.h"
@@ -47,9 +48,8 @@ class NET_EXPORT GZipHeader {
   // gzip header, return INVALID_HEADER. When we've seen a complete
   // gzip header, return COMPLETE_HEADER and set the pointer pointed
   // to by header_end to the first byte beyond the gzip header.
-  Status ReadMore(const char* inbuf,
-                  int inbuf_len,
-                  const char** header_end);
+  Status ReadMore(const char* inbuf, size_t inbuf_len, const char** header_end);
+
  private:
   enum {                       // flags (see RFC)
     FLAG_FTEXT        = 0x01,  // bit 0 set: file probably ascii text
